fabricio - solved issues
