+++ SYL_project/settings.py
@@ -31,7 +31,7 @@ ALLOWED_HOSTS = []
 # Application definition
 
 # Le indica a Django que va a haber una app con este nombre
-CUSTOM_APPS = ["syl_app"]
+CUSTOM_APPS = ["app_viajes"]
 
 INSTALLED_APPS = [
     "django.contrib.admin",
@@ -125,7 +125,7 @@ USE_TZ = True
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/4.2/howto/static-files/
 
-STATIC_URL = 'static/'
+STATIC_URL = "static/"
 
 STATICFILES_DIRS = [BASE_DIR / "static"]
