--- app_viajes/Templates/viajes_delete.html
@@ -1,16 +1,0 @@
-{% extends 'base_viajes.html' %} {% block content %}
-
-<main>
-    <div class="main-color-viajes">
-        <h1>Estas seguro ue deseas borrar este viaje?</h1>
-        <h2>{{base_viajes.nombre}}</h2>
-        <form method="post">
-            {% csrf_token %}
-            <div>{{form.as_p}}</div>
-            <input type="submit" value="{{ tipo }}" />
-        </form>
-
-        <a href="{% url 'viajes:all' %}">Cancelar</a>
-    </div>
-</main>
-{% endblock %}
