--- syl_app/templates/cruceros_create.html
@@ -1,27 +1,0 @@
-{% extends 'index.html' %} {% load static %}
-
-{% block content %}
-
-<section class="viajes_container">
-  <h2>CREAR VIAJE</h2>
-  <div class="viajes_create">
-    <form class="form_app" method="post">
-      {% csrf_token %}
-
-      {% for field in form %}
-      <div class="form-field">
-        {{ field.label_tag }}
-        {{ field }}
-        {{ field.errors }}
-      </div>
-      {% endfor %}
-      <input class="formulario-form-button a_button" type="submit" value="{{ tipo }}" />
-    </form>
-  </div>
-  
-  <div class="btn">
-    <a class="button1" href="{% url 'cruceros:all' %}">Volver</a>
-  </div>
-</section>
-
-{% endblock %}
