--- syl_app/templates/cruceros_detail.html
@@ -1,26 +1,0 @@
-{% extends 'index.html' %}
-{% load static %}
-
-{% block content %}
-
-<section class="viajes_container">
-  <div class="contenedor_detail">
-    <div class="contenedor_item">
-      <img class="contenedor_imagen"
-        src="{% if compania_imagen %}{{ compania_imagen }}{% else %}{% static 'img/default.jpg' %}{% endif %}"
-        alt="Imagen de Compañía">
-    </div>
-  </div>
-
-  <div class="txt_detail">
-    <h5 class="">Empresa: {{ cruceros.compania }}</h5>
-    <h5 class=""> Viaja a <u>{{ cruceros.continente }}</u></h5>
-    <h5 class=""> Desde {{ cruceros.puerto }}</h5>
-    <h5 class="">Valor base: <strong>{{ cruceros.precio }}</strong></h5>
-  </div>
-  <div class="btn">
-    <a class="button1" href="{% url 'cruceros:all' %}">Volver</a>
-  </div>
-</section>
-
-{% endblock %}
