--- templates/base_viajes.html
@@ -1,36 +1,0 @@
-{% load static %}
-
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="{% static './css/main.css' %}" />
-    <title>Document</title>
-</head>
-
-<body>
-    <div id="app">
-        <header id="header">
-            <ul>
-                <li v-for="item in lista" @click="clickedShow">
-                    <a v-bind:href="`#${item.toLowerCase()}`">-- item --</a>
-                </li>
-                <li>
-                    <a href="{% url 'viajes:all' %}">viajes</a>
-                </li>
-            </ul>
-        </header>
-
-        {% block content %}
-        {% endblock %}
-
-    </div>
-    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-    <script src="https://unpkg.com/vue@next"></script>
-    <script src="{% static './js/vueScript.js' %}"></script>
-</body>
-
-</html>
